en: 实时校验推文改为空投前校验
