+++ internal/job/cmd/config.demo.yaml
@@ -47,6 +47,8 @@ contract:
 # twitter configuration
 twitter:
   bear-token: ""
+  claim-content: "我通过了 @DecertMe 上完成了一个挑战并获得了链上能力认证的徽章。\nhttps://decert.me/quests/\n#DecertMe"
+
 
 # scheduler configuration
 scheduler:
